requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyio==3.7.0
    # via httpcore
asgiref==3.7.2
    # via flask
asyncio-throttle==1.0.2
    # via -r requirements.in
black==23.3.0
    # via -r requirements.in
blinker==1.6.2
    # via flask
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   flask
flake8==6.0.0
    # via -r requirements.in
flask[async]==2.3.1
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==0.17.2
    # via httpx
httpx==0.24.1
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
isort==5.12.0
    # via -r requirements.in
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via black
pathspec==0.11.1
    # via black
platformdirs==3.8.0
    # via black
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
urllib3==2.0.3
    # via requests
werkzeug==2.3.6
    # via flask
